set secure permissions on temporary file
Ensure the log files that are used to temporarily hold the output
of the wkhtmltopdf binary have proper permissions set. Apparently,
Java creates temporary files that are word-readable by default
and Oracle appears to be unwilling to fix this major security flaw.

The log file is now also properly removed.